pd(migrate): refactor/begin forming an interface for migrations
