+++ crates/bin/pd/src/migrate.rs
@@ -4,21 +4,19 @@
 //! node operators must coordinate to perform a chain upgrade.
 //! This module declares how local `pd` state should be altered, if at all,
 //! in order to be compatible with the network post-chain-upgrade.
+mod reset_halt_bit;
+mod simple;
 mod testnet72;
 mod testnet74;
 
 use anyhow::{ensure, Context};
-use futures::StreamExt as _;
-use penumbra_governance::{StateReadExt, StateWriteExt};
+use penumbra_governance::StateReadExt;
+use penumbra_sct::component::clock::EpochRead;
 use std::path::{Path, PathBuf};
 use tracing::instrument;
 
-use cnidarium::{StateDelta, StateRead, StateWrite, Storage};
-use jmt::RootHash;
-use penumbra_app::{app::StateReadExt as _, SUBSTORE_PREFIXES};
-use penumbra_sct::component::clock::{EpochManager, EpochRead};
-
-use crate::testnet::generate::TestnetConfig;
+use cnidarium::Storage;
+use penumbra_app::SUBSTORE_PREFIXES;
 
 use flate2::write::GzEncoder;
 use flate2::Compression;
@@ -32,8 +30,6 @@ pub enum Migration {
     /// A simple migration: adds a key to the consensus state.
     /// This is useful for testing upgrade mechanisms, including in production.
     SimpleMigration,
-    /// Testnet-70 migration: move swap executions from the jmt to nv-storage.
-    Testnet70,
     /// Testnet-72 migration:
     /// - Migrate `BatchSwapOutputData` to new protobuf, replacing epoch height with index.
     Testnet72,
@@ -45,30 +41,30 @@ pub enum Migration {
 }
 
 impl Migration {
-    #[instrument(skip(path_to_export, genesis_start, force))]
+    #[instrument(skip(pd_home, genesis_start, force))]
     pub async fn migrate(
         &self,
-        path_to_export: PathBuf,
+        pd_home: PathBuf,
+        comet_home: Option<PathBuf>,
         genesis_start: Option<tendermint::time::Time>,
         force: bool,
     ) -> anyhow::Result<()> {
         tracing::debug!(
-            ?path_to_export,
+            ?pd_home,
             ?genesis_start,
             ?force,
             "preparing to run migration!"
         );
-        let rocksdb_dir = path_to_export.join("rocksdb");
+        let rocksdb_dir = pd_home.join("rocksdb");
         let storage = Storage::load(rocksdb_dir, SUBSTORE_PREFIXES.to_vec()).await?;
         let export_state = storage.latest_snapshot();
         let root_hash = export_state.root_hash().await.expect("can get root hash");
-        let app_hash_pre_migration: RootHash = root_hash.into();
         let height = export_state
             .get_block_height()
             .await
             .expect("can get block height");
 
-        tracing::debug!(?app_hash_pre_migration, current_height = ?height, ?force,
+        tracing::debug!(pre_migration_root_hash = ?root_hash, current_height = ?height, ?force,
             "determining if the chain is halted and the migration is allowed to run");
 
         ensure!(
@@ -76,167 +72,30 @@ impl Migration {
             "to run a migration, the chain halt bit must be set to `true` or use the `--force` cli flag"
         );
 
-        tracing::info!(?app_hash_pre_migration, pre_upgrade_height = ?height, ?self, "started migration");
+        tracing::info!(pre_migration_root_hash = ?root_hash, pre_upgrade_height = ?height, ?self, "started migration");
         match self {
             Migration::ReadyToStart => {
-                let mut delta = StateDelta::new(export_state);
-                delta.ready_to_start();
-                let _ = storage.commit_in_place(delta).await?;
-                tracing::info!(
-                    "migration completed: halt bit is turned off, chain is ready to start"
-                );
-                Ok(())
+                reset_halt_bit::migrate(storage, pd_home, genesis_start).await?;
+                return Ok(());
             }
             Migration::SimpleMigration => {
-                let post_ugprade_height = height.wrapping_add(1);
-
-                /* --------- writing to the jmt  ------------ */
-                tracing::info!(?app_hash_pre_migration, "app hash pre-upgrade");
-                let mut delta = StateDelta::new(export_state);
-                delta.put_raw(
-                    "banana".to_string(),
-                    "a good fruit (and migration works!)".into(),
-                );
-                delta.put_block_height(0u64);
-                let root_hash = storage.commit_in_place(delta).await?;
-                let app_hash_post_migration: RootHash = root_hash.into();
-                tracing::info!(?app_hash_post_migration, "app hash post upgrade");
-
-                /* --------- collecting genesis data -------- */
-                tracing::info!("generating genesis");
-                let migrated_state = storage.latest_snapshot();
-                let root_hash = migrated_state.root_hash().await.expect("can get root hash");
-                let app_hash: RootHash = root_hash.into();
-                tracing::info!(?root_hash, "root hash from snapshot (post-upgrade)");
-
-                /* ---------- generate genesis ------------  */
-                let chain_id = migrated_state.get_chain_id().await?;
-                let app_state = penumbra_app::genesis::Content {
-                    chain_id,
-                    ..Default::default()
-                };
-                let mut genesis =
-                    TestnetConfig::make_genesis(app_state.clone()).expect("can make genesis");
-                genesis.app_hash = app_hash
-                    .0
-                    .to_vec()
-                    .try_into()
-                    .expect("infaillible conversion");
-                genesis.initial_height = post_ugprade_height as i64;
-                genesis.genesis_time = genesis_start.unwrap_or_else(|| {
-                    let now = tendermint::time::Time::now();
-                    tracing::info!(%now, "no genesis time provided, detecting a testing setup");
-                    now
-                });
-                let checkpoint = app_hash.0.to_vec();
-                let genesis = TestnetConfig::make_checkpoint(genesis, Some(checkpoint));
-
-                let genesis_json = serde_json::to_string(&genesis).expect("can serialize genesis");
-                tracing::info!("genesis: {}", genesis_json);
-                let genesis_path = path_to_export.join("genesis.json");
-                std::fs::write(genesis_path, genesis_json).expect("can write genesis");
-
-                let validator_state_path = path_to_export.join("priv_validator_state.json");
-                let fresh_validator_state =
-                    crate::testnet::generate::TestnetValidator::initial_state();
-                std::fs::write(validator_state_path, fresh_validator_state)
-                    .expect("can write validator state");
-                Ok(())
+                simple::migrate(storage, pd_home.clone(), genesis_start).await?
             }
-            Migration::Testnet70 => {
-                // Our goal is to fetch all swap executions from the jmt and store them in nv-storage.
-                // In particular, we want to make sure that client lookups for (height, trading pair)
-                // resolve to the same value as before.
-
-                // Setup:
-                let start_time = std::time::SystemTime::now();
-                let rocksdb_dir = path_to_export.join("rocksdb");
-                let storage =
-                    Storage::load(rocksdb_dir.clone(), SUBSTORE_PREFIXES.to_vec()).await?;
-                let export_state = storage.latest_snapshot();
-                let root_hash = export_state.root_hash().await.expect("can get root hash");
-                let pre_upgrade_root_hash: RootHash = root_hash.into();
-                let pre_upgrade_height = export_state
-                    .get_block_height()
-                    .await
-                    .expect("can get block height");
-                let post_upgrade_height = pre_upgrade_height.wrapping_add(1);
-
-                // We initialize a `StateDelta` and start by reaching into the JMT for all entries matching the
-                // swap execution prefix. Then, we write each entry to the nv-storage.
-                let mut delta = StateDelta::new(export_state);
-
-                let prefix_key = "dex/swap_execution/";
-                let mut swap_execution_stream = delta.prefix_raw(prefix_key);
-
-                while let Some(r) = swap_execution_stream.next().await {
-                    let (key, swap_execution) = r?;
-                    tracing::info!("migrating swap execution: {}", key);
-                    delta.nonverifiable_put_raw(key.into_bytes(), swap_execution);
-                }
-
-                delta.put_block_height(0u64);
-
-                let post_upgrade_root_hash = storage.commit_in_place(delta).await?;
-                tracing::info!(?post_upgrade_root_hash, "post-upgrade root hash");
-
-                let migration_duration = start_time.elapsed().expect("start time not set");
-
-                // Reload storage so we can make reads against its migrated state:
-                storage.release().await;
-                let storage = Storage::load(rocksdb_dir, SUBSTORE_PREFIXES.to_vec()).await?;
-                let migrated_state = storage.latest_snapshot();
-                storage.release().await;
-
-                // The migration is complete, now we need to generate a genesis file. To do this, we need
-                // to lookup a validator view from the chain, and specify the post-upgrade app hash and
-                // initial height.
-                let chain_id = migrated_state.get_chain_id().await?;
-                let app_state = penumbra_app::genesis::Content {
-                    chain_id,
-                    ..Default::default()
-                };
-                let mut genesis =
-                    TestnetConfig::make_genesis(app_state.clone()).expect("can make genesis");
-                genesis.app_hash = post_upgrade_root_hash
-                    .0
-                    .to_vec()
-                    .try_into()
-                    .expect("infaillible conversion");
-                genesis.initial_height = post_upgrade_height as i64;
-                genesis.genesis_time = genesis_start.unwrap_or_else(|| {
-                    let now = tendermint::time::Time::now();
-                    tracing::info!(%now, "no genesis time provided, detecting a testing setup");
-                    now
-                });
-                let checkpoint = post_upgrade_root_hash.0.to_vec();
-                let genesis = TestnetConfig::make_checkpoint(genesis, Some(checkpoint));
-
-                let genesis_json = serde_json::to_string(&genesis).expect("can serialize genesis");
-                tracing::info!("genesis: {}", genesis_json);
-                let genesis_path = path_to_export.join("genesis.json");
-                std::fs::write(genesis_path, genesis_json).expect("can write genesis");
-
-                let validator_state_path = path_to_export.join("priv_validator_state.json");
-                let fresh_validator_state =
-                    crate::testnet::generate::TestnetValidator::initial_state();
-                std::fs::write(validator_state_path, fresh_validator_state)
-                    .expect("can write validator state");
-
-                tracing::info!(
-                    pre_upgrade_height,
-                    post_upgrade_height,
-                    ?pre_upgrade_root_hash,
-                    ?post_upgrade_root_hash,
-                    duration = migration_duration.as_secs(),
-                    "successful migration!"
-                );
-
-                Ok(())
+            Migration::Testnet72 => {
+                testnet72::migrate(storage, pd_home.clone(), genesis_start).await?
             }
-            Migration::Testnet72 => testnet72::migrate(path_to_export, genesis_start).await,
-            Migration::Testnet74 => testnet74::migrate(path_to_export, genesis_start).await,
+            Migration::Testnet74 => {
+                testnet74::migrate(storage, pd_home.clone(), genesis_start).await?
+            }
+        };
+
+        if let Some(comet_home) = comet_home {
+            // TODO avoid this when refactoring to clean up migrations
+            let genesis_path = pd_home.join("genesis.json");
+            migrate_comet_data(comet_home, genesis_path).await?;
         }
+
+        Ok(())
     }
 }
